+++ src/base/FormControl/FormControl.styled.ts
@@ -1,0 +1,27 @@
+/* Vendors */
+import styled from 'styled-components'
+
+/* Types */
+
+/* Utils */
+
+/* Styles */
+export const FormLabel = styled.label`
+  font-family: var(--font-sans);
+  font-size: var(--font-size);
+  font-weight: bold;
+  margin-bottom: 8px;
+  display: block;
+  line-height: 1;
+`
+
+export const FormCaption = styled.span`
+  font-family: var(--font-sans);
+  font-size: 14px;
+  color: var(--color-grey);
+  display: block;
+  line-height: 1;
+  margin-top: 2px;
+`
+
+export const FormControl = styled.div``