+++ src/base/FormControl/FormControl.types.ts
@@ -1,0 +1,10 @@
+/* Vendor types */
+import {ReactElement} from 'react'
+
+/* Types */
+export interface Props {
+  className?: string,
+  label?: string,
+  caption?: string,
+  children: ReactElement,
+}